--- hrs-portlets-api/src/main/java/edu/byu/hr/HrPortletRuntimeException.java
@@ -1,41 +1,0 @@
-/**
- * Licensed to Jasig under one or more contributor license
- * agreements. See the NOTICE file distributed with this work
- * for additional information regarding copyright ownership.
- * Jasig licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file
- * except in compliance with the License. You may obtain a
- * copy of the License at:
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package edu.byu.hr;
-
-/**
- * HR Portlets Runtime Exception
- *
- * @author James Wennmacher, jwennmacher@unicon.net
- */
-
-public class HrPortletRuntimeException extends RuntimeException {
-
-    public HrPortletRuntimeException(String message) {
-        super(message);
-    }
-
-    public HrPortletRuntimeException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public HrPortletRuntimeException(Throwable cause) {
-        super(cause);
-    }
-}
