+++ hrs-portlets-employee-time-reporting/pom.xml
@@ -111,14 +111,6 @@
             <artifactId>joda-time</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.codehaus.jackson</groupId>
-            <artifactId>jackson-core-asl</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.codehaus.jackson</groupId>
-            <artifactId>jackson-mapper-asl</artifactId>
-        </dependency>
-        <dependency>
             <groupId>org.jasig.portlet.datalist</groupId>
             <artifactId>datalist-portlet-utils</artifactId>
         </dependency>
