--- hrs-portlets-employee-time-reporting/src/main/java/edu/byu/hr/timereporting/util/TimePunchEmployeeKeyGenerator.java
@@ -1,37 +1,0 @@
-/**
- * Licensed to Jasig under one or more contributor license
- * agreements. See the NOTICE file distributed with this work
- * for additional information regarding copyright ownership.
- * Jasig licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file
- * except in compliance with the License. You may obtain a
- * copy of the License at:
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package edu.byu.hr.timereporting.util;
-
-/**
- * Returns a cache key for the time punch entries
- *
- * @author James Wennmacher, jwennmacher@unicon.net
- */
-
-public interface TimePunchEmployeeKeyGenerator {
-
-    /**
-     * Returns a cache key for the user
-     *
-     * @param emplId user's employeeId
-     * @return cache key for the user
-     */
-    String getKey(String emplId);
-}
