+++ hrs-portlets-employee-time-reporting/src/main/resources/app-context/timereportingAppContext.xml
@@ -32,7 +32,7 @@
         http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.1.xsd http://www.springframework.org/schema/cache http://www.springframework.org/schema/cache/spring-cache.xsd">
 
     <context:annotation-config />
-    <context:component-scan base-package="edu.byu.hr.timereporting" />
+    <context:component-scan base-package="org.apereo.portlet.hr.timereporting" />
     <cache:annotation-driven cache-manager="springCacheManager"/>
 
     <context:property-override location="classpath:/timereporting-override.properties" />
